requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
google-generativeai>=0.8.6
anthropic==0.18.1
redis==5.0.1
python-multipart==0.0.6
pypdf2==3.0.1
python-dotenv==1.0.0
sse-starlette==1.8.2
httpx==0.26.0
